Refactor classes.
