--- packages/core/src/code_assist/constants.ts
@@ -1,8 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export const CCPA_ENDPOINT = 'https://cloudcode-pa.googleapis.com';
-export const CCPA_API_VERSION = '/v1internal';
