--- packages/core/src/code_assist/metadata.ts
@@ -1,37 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export interface ClientMetadata {
-  ideType?: ClientMetadataIdeType | null;
-  ideVersion?: string | null;
-  pluginVersion?: string | null;
-  platform?: ClientMetadataPlatform | null;
-  updateChannel?: string | null;
-  duetProject?: string | null;
-  pluginType?: ClientMetadataPluginType | null;
-  ideName?: string | null;
-}
-
-export type ClientMetadataIdeType =
-  | 'IDE_UNSPECIFIED'
-  | 'VSCODE'
-  | 'INTELLIJ'
-  | 'VSCODE_CLOUD_WORKSTATION'
-  | 'INTELLIJ_CLOUD_WORKSTATION'
-  | 'CLOUD_SHELL';
-export type ClientMetadataPlatform =
-  | 'PLATFORM_UNSPECIFIED'
-  | 'DARWIN_AMD64'
-  | 'DARWIN_ARM64'
-  | 'LINUX_AMD64'
-  | 'LINUX_ARM64'
-  | 'WINDOWS_AMD64';
-export type ClientMetadataPluginType =
-  | 'PLUGIN_UNSPECIFIED'
-  | 'CLOUD_CODE'
-  | 'GEMINI'
-  | 'AIPLUGIN_INTELLIJ'
-  | 'AIPLUGIN_STUDIO';
